pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "handwrite-font-maker"
version = "0.1.0"
description = "Convert handwriting specimen sheets into installable fonts."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26",
  "Pillow>=10.0",
  "opencv-python-headless>=4.13,<5",
  "reportlab>=4.0",
  "qrcode[pil]>=8.0",
]

[project.scripts]
handwrite-font-maker = "handwrite_font_maker.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
test = [
  "pytest>=8.0",
]
